% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-braid.R, R/stat-braid.R
\name{geom_braid}
\alias{geom_braid}
\alias{stat_braid}
\title{Braids}
\usage{
geom_braid(
  mapping = NULL,
  data = NULL,
  stat = "braid",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)

stat_braid(
  mapping = NULL,
  data = NULL,
  geom = "braid",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\description{
`geom_braid()` is an extension of `geom_ribbon()` to
appropriately fill the area between two alternating lines
with different colors, whether one line is over or under
the other.
}
\seealso{
[geom_ribbon()] for non-alternating lines
}
