% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbraid.R
\docType{package}
\name{ggbraid-package}
\alias{ggbraid}
\alias{ggbraid-package}
\title{ggbraid: Braid Ribbons in ggplot2}
\description{
ggbraid provides a new stat, `stat_braid()`, that extends the functionality of `geom_ribbon()` to correctly fill the area between two alternating lines (or steps) with two different colors. ggbraid also provides a geom, `geom_braid()`, that wraps `geom_ribbon()` and uses `stat_braid()` by default.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nsgrantham.github.io/ggbraid}
  \item \url{https://github.com/nsgrantham/ggbraid}
}

}
\keyword{internal}
