% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbraid.R
\docType{package}
\name{ggbraid-package}
\alias{ggbraid}
\alias{ggbraid-package}
\title{ggbraid: Fill the area between two alternating lines with `geom_braid()`}
\description{
It is difficult to properly fill the area between two lines in `ggplot2`. Use of `geom_ribbon()` is encouraged, but it does not always work as expected due to a potentional mismatch between the data and how `geom_ribbon()` requires the data to be arranged. This package introduces `geom_braid()`, a substitute for `geom_ribbon()` that takes care of these details before applying `geom_ribbon()`, resulting in a correctly filled area between two alternating lines.
}
\keyword{internal}
